Return non-zero exit code breaks first call of systemd unit
When installing the anything-sync-daemon and run  "systemctl start"
before adjust asd.conf, the systemd unit returns it is failed.
This is caused by the return of a non-zero exit code if no directory
to sync is set or the $VOLATILE is not adjusted yet.

This commit sets a zero exit code for this not uncommon cases so the
call of the "systemctl start" does not fail but just stop due to the
exit command.

This is necessary since so package manager like dpkg can try to start
the systemd unit after installation/upgrade and do not fail due to a
failure of the systemd unit.